pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "snipgen"
version = "0.1.0"
description = "AI-driven CRISPR guide RNA design tool"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "biopython>=1.83",
    "pandas>=2.0",
    "numpy>=1.26",
    "click>=8.1",
    "tabulate>=0.9",
    "joblib>=1.3",
    "scikit-learn>=1.3",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov", "mypy", "ruff"]
ml  = ["torch>=2.0", "onnxruntime>=1.17"]

[project.scripts]
snipgen = "snipgen.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=snipgen --cov-report=term-missing"

[tool.ruff]
line-length = 100
target-version = "py311"
